feat(client): `StatelessL2BlockExecutor` (#210)
* feat(client): `StatelessL2BlockExecutor`

Implements the `StatelessL2BlockExecutor`, capable of executing
`L2PayloadAttributes` and returning the block header + receipts.

The `StatelessL2BlockExecutor` is backed by the `TrieDB` in the
`kona-mpt` crate, fetching necessary data from the initial state root as
well as the starting parent block hash.

* fix(client): Trie fixes, etc.

x

trim test files

* big block testin'

* checkpoint: big block works

* checkpoint

dep reordering

* cleanups

* more tests

+++ crates/primitives/src/block_info.rs
@@ -145,7 +145,7 @@ impl L1BlockInfoTx {
             Ok(Self::Ecotone(L1BlockInfoEcotone {
                 number: l1_header.number,
                 time: l1_header.timestamp,
-                base_fee: l1_header.base_fee_per_gas.unwrap_or(0),
+                base_fee: l1_header.base_fee_per_gas.unwrap_or(0) as u64,
                 block_hash: l1_header.hash_slow(),
                 sequence_number,
                 batcher_address: system_config.batcher_addr,
@@ -157,7 +157,7 @@ impl L1BlockInfoTx {
             Ok(Self::Bedrock(L1BlockInfoBedrock {
                 number: l1_header.number,
                 time: l1_header.timestamp,
-                base_fee: l1_header.base_fee_per_gas.unwrap_or(0),
+                base_fee: l1_header.base_fee_per_gas.unwrap_or(0) as u64,
                 block_hash: l1_header.hash_slow(),
                 sequence_number,
                 batcher_address: system_config.batcher_addr,
@@ -469,7 +469,7 @@ mod test {
 
         assert_eq!(l1_info.number, l1_header.number);
         assert_eq!(l1_info.time, l1_header.timestamp);
-        assert_eq!(l1_info.base_fee, l1_header.base_fee_per_gas.unwrap_or(0));
+        assert_eq!(l1_info.base_fee, l1_header.base_fee_per_gas.unwrap_or(0) as u64);
         assert_eq!(l1_info.block_hash, l1_header.hash_slow());
         assert_eq!(l1_info.sequence_number, sequence_number);
         assert_eq!(l1_info.batcher_address, system_config.batcher_addr);
@@ -500,7 +500,7 @@ mod test {
 
         assert_eq!(l1_info.number, l1_header.number);
         assert_eq!(l1_info.time, l1_header.timestamp);
-        assert_eq!(l1_info.base_fee, l1_header.base_fee_per_gas.unwrap_or(0));
+        assert_eq!(l1_info.base_fee, l1_header.base_fee_per_gas.unwrap_or(0) as u64);
         assert_eq!(l1_info.block_hash, l1_header.hash_slow());
         assert_eq!(l1_info.sequence_number, sequence_number);
         assert_eq!(l1_info.batcher_address, system_config.batcher_addr);
